move class into single files and update imports in router
