+++ api/router.py
@@ -1,11 +1,28 @@
 import json
-from fastapi import APIRouter, HTTPException, status
-from service.walletService import PrivateKeyService
-from shemas.shemas import *
-from core.config import contract_settings
-from web3.middleware import geth_poa_middleware
-from web3 import Web3
 from typing import List, Optional
+
+from fastapi import APIRouter, HTTPException, status
+from web3 import Web3
+from web3.middleware import geth_poa_middleware
+
+from core.config import contract_settings
+from service.walletService import PrivateKeyService
+from shemas.CreateWalletResponse import CreateWalletResponse
+from shemas.Data import Data
+from shemas.MetaData import MetaData
+from shemas.MetaDataRequest import MetaDataRequest
+from shemas.MetaDataResponse import MetaDataResponse
+from shemas.MintOneToManyRessourceRequest import MintOneToManyRessourceRequest
+from shemas.MintRessourceRequest import MintRessourceRequest
+from shemas.ResourceCreatedEventResponse import ResourceCreatedEventResponse
+from shemas.ResourceMetaDataChangedEventResponse import ResourceMetaDataChangedEventResponse
+from shemas.ResourceTemplateResponse import ResourceTemplateResponse
+from shemas.RoleAssignmentRequest import RoleAssignmentRequest
+from shemas.RoleResponse import RoleResponse
+from shemas.TransactionRequest import TransactionRequest
+from shemas.TransactionResponse import TransactionResponse
+from shemas.TransferResourceRequest import TransferResourceRequest
+from shemas.TransferResourceResponse import TransferResourceResponse
 
 web3 = Web3(Web3.HTTPProvider(contract_settings.validator_address))
 web3.middleware_onion.inject(geth_poa_middleware, layer=0)
