+++ shemas/CreateWalletResponse.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class CreateWalletResponse(BaseModel):
+    wallet_address: str
