+++ shemas/EventResponse.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+from typing import List
+
+class EventResponse(BaseModel):
+    event: str
+    data: List[dict]
