+++ shemas/MintOneToManyRessourceRequest.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class MintOneToManyRessourceRequest(BaseModel):
+    from_wallet_address: str
+    producer_token_id: int
+    metaData: dict
