+++ shemas/MintToManyDataRequest.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class MintToManyDataRequest(BaseModel):
+    tokenId: int
+    metaData: dict
