+++ shemas/ResourceCreatedEventResponse.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class ResourceCreatedEventResponse(BaseModel):
+    tokenId: int
+    ressourceName: str
+    message: str
+    caller: str
