+++ shemas/RoleAssignmentRequest.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class RoleAssignmentRequest(BaseModel):
+    from_wallet_address: str
+    target_wallet_address: str
+    role: str
