+++ shemas/RoleResponse.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class RoleResponse(BaseModel):
+    role: str
