+++ shemas/TransactionRequest.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+class TransactionRequest(BaseModel):
+    sender_address: str
+    receiver_address: str
+    amount: float  # Amount in Ether
