+++ shemas/TransactionResponse.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class TransactionResponse(BaseModel):
+    transaction_hash: str
+    details: str
