+++ shemas/TransferResourceResponse.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+class TransferResourceResponse(BaseModel):
+    tokenId: int
+    quantity: int
+    from_wallet_address: str
+    to_wallet_address: str
